LZMR-47 added request mapping for characters controller
